// data/github/current/ui/packages/item-picker/components/__generated__/OrganizationListPaginationQuery.graphql.ts
export type OrganizationListPaginationQuery$variables = {
  count?: number | null | undefined;
  cursor?: string | null | undefined;
  id: string;
};
export type OrganizationListPaginationQuery$data = {
  readonly node: {
    readonly " $fragmentSpreads": FragmentRefs<"OrganizationPickerOrganizationsPaginated">;
  } | null | undefined;
};
export type OrganizationListPaginationQuery = {
  response: OrganizationListPaginationQuery$data;
  variables: OrganizationListPaginationQuery$variables;
};