// data/github/current/ui/packages/item-picker/components/__generated__/OrganizationPickerOrganizationsPaginated.graphql.ts
export type OrganizationPickerOrganizationsPaginated$data = {
  readonly id: string;
  readonly organizations: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly " $fragmentSpreads": FragmentRefs<"OrganizationPickerOrganization">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "OrganizationPickerOrganizationsPaginated";
};
export type OrganizationPickerOrganizationsPaginated$key = {
  readonly " $data"?: OrganizationPickerOrganizationsPaginated$data;
  readonly " $fragmentSpreads": FragmentRefs<"OrganizationPickerOrganizationsPaginated">;
};