// data/github/current/ui/packages/item-picker/components/__generated__/OrganizationPickerQuery.graphql.ts
export type OrganizationPickerQuery$variables = Record<PropertyKey, never>;
export type OrganizationPickerQuery$data = {
  readonly viewer: {
    readonly " $fragmentSpreads": FragmentRefs<"OrganizationPickerOrganizationsPaginated">;
  };
};
export type OrganizationPickerQuery = {
  response: OrganizationPickerQuery$data;
  variables: OrganizationPickerQuery$variables;
};